Перенос функции рендеринга из src/utils.js в src/utils/render.js
